perf(mappings): avoid computing error string on hot path
Profiling with perfanno.nvim shows that a large portion of startup time
was spent on this line due to the computation of the error string.

This commit avoids computing the problematic error string unless it is
going to be raised.

Startup time as recorded by lazy.nvim shows a reduction of 28% with this
change: 13.43ms -> 10.49ms.