refactor(util): simplify parse_internal
Rewrites parse_internal as a state machine. This reduces the number of
variables required to manage the parser, making it simpler to
understand.

The code is optimized with LuaJIT in mind, in particular:

- Make use of composite conditionals on hot path to leverage IR
  optimizations.

- Avoid string building which has quadratic complexity in Lua.

Startup time goes from 10.1ms[+-0.13] down to 9.81ms[+-0.165] (~2.8%
improvement) with this commit as recorded by lazy.nvim with samples of
10.