fix(util): only collect valid <> keys (#438)
neovim is very lenient on the parsing of `<` in mappings. If the `<>`
pairs did not enclose a valid special key, the entire sequence will be
treated as `<lt>[keys here]>`.

To handle this, simply cross reference against the internal
representation to see if the key should be treated as a special key.