keys: Magically detect key groups
from nvim-builtin keymappings. Scans for a magic string in the `desc`
field of a keymapping and, if found, removes the magic prefix and
registers the respective binding as a group, rather than an action.

This allows defining keybinding groups from outside `which-key`, which
can be utilized in conjunction with e.g. the `keys` config option for
the `lazy.nvim` plugin manager.