fix(state): schedule modechanged event handling. Fixes #735
When enabling vim.opt.keymodel = 'startsel' and using the key
shift+motion to start a selection (for example shift+Down), the screen
is not updated properly with the visual selection until after the
ModeChanged event is completed. Re-scheduling the which-key.start()
inside the ModeChanged event handler will update the screen properly.