fix(registers): improve detection of osc52 registers
The prior implementation assumed that registers "+" and "*" used OSC52
if and only if vim.g.clipboard.name = "OSC 52". Doing so is incorrect:
a user may configure a custom clipboard provider with a different name,
or set a different paste source for "+" and "*", or even set the option
to a string as shorthand, i.e., vim.g.clipboard = 'osc52'.

This implementation performs more robust checking to ensure we don't try
to read from OSC52 registers by discriminating vim.g.clipboard more
closely.